Reverted use of Immutables
Haven't been able to get Visual Studio Code to recognize the
generated Immutables implementations of classes even though
they are visible in the
build/generated/sources/annotationProcessor/java/main/
directory, and time-boxing this after a few hours of trial
and error.

May revisit in the future but this is an aside to the
purpose of the project, so will continue with normal
boilerplate code for the time being.